chore: move bls verifier to separate module
