+++ modules/light-clients/08-wasm/blsverifier/crypto.go
@@ -1,4 +1,4 @@
-package customquery
+package blsverifier
 
 import (
 	"fmt"
